Allow points to be passed by ownership or by ref
`write_points` expects an owned `Iterator` of `Point`. It would be more
convienient to allow the points to be passed by reference or by
ownership.

impl `IntoIterator` for `&Points` and update trait bounds on
`write_points` and a few others to `IntoIterator<Item = impl Borrow<Point>>`

This allows the user to pass either by ownership or by reference.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>