Use `Cow` in `Value` for String
When adding tags and field values, the `value` param is converted to a `String` if it's a `&str`.

We can optimize this case by internally switching  `Value::String` to contain a `Cow<'a, str>`.

In addition, update the `line_serialization` fn to push into a `mut String` rather than a `Vec`.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>